fix: 解决重连失效的问题，当时为了调度到异常的sql导致
